Moved the user input validations into its own file
