+++ brambl-sdk/src/main/scala/co/topl/brambl/builders/BuilderError.scala
@@ -7,3 +7,7 @@ package co.topl.brambl.builders
  * @param message The error message
  */
 abstract class BuilderError(message: String, cause: Throwable = null) extends RuntimeException(message, cause)
+
+case class UserInputError(message: String) extends BuilderError(message, null)
+
+case class BuilderRuntimeError(message: String, cause: Throwable = null) extends BuilderError(message, cause)
