+++ brambl-sdk/src/main/scala/co/topl/brambl/builders/UserInputValidations.scala
@@ -1,0 +1,163 @@
+package co.topl.brambl.builders
+
+import cats.data.{Chain, NonEmptyChain, Validated, ValidatedNec}
+import cats.implicits.{catsSyntaxOptionId, catsSyntaxValidatedIdBinCompat0, toFoldableOps}
+import co.topl.brambl.models.box.{AssetMintingStatement, FungibilityType}
+import co.topl.brambl.models.{LockAddress, SeriesId, TransactionOutputAddress}
+import co.topl.brambl.models.box.Value._
+import quivr.models.Int128
+import co.topl.brambl.syntax.int128AsBigInt
+import co.topl.genus.services.Txo
+
+object UserInputValidations {
+
+  def txoAddressMatch(
+    testAddr:      TransactionOutputAddress,
+    expectedAddr:  TransactionOutputAddress,
+    testLabel:     String,
+    expectedLabel: String
+  ): ValidatedNec[UserInputError, Unit] =
+    Validated.condNec(testAddr == expectedAddr, (), UserInputError(s"$testLabel does not match $expectedLabel"))
+
+  def isLvls(testValue: Value, testLabel: String): ValidatedNec[UserInputError, Unit] =
+    Validated.condNec(testValue.isLvl, (), UserInputError(s"$testLabel does not contain LVLs"))
+
+  def isGroup(testValue: Value, testLabel: String): ValidatedNec[UserInputError, Unit] =
+    Validated.condNec(testValue.isGroup, (), UserInputError(s"$testLabel does not contain Group Constructor Tokens"))
+
+  def isSeries(testValue: Value, testLabel: String): ValidatedNec[UserInputError, Unit] =
+    Validated.condNec(
+      testValue.isSeries,
+      (),
+      UserInputError(s"$testLabel does not contain Series Constructor Tokens")
+    )
+
+  def fixedSeriesMatch(
+    testValue:     Option[SeriesId],
+    expectedValue: Option[SeriesId]
+  ): ValidatedNec[UserInputError, Unit] =
+    (testValue, expectedValue) match {
+      case (Some(fixedSeries), Some(expectedSeries)) =>
+        Validated.condNec(
+          fixedSeries == expectedSeries,
+          (),
+          UserInputError(s"fixedSeries does not match provided Series ID")
+        )
+      case _ => ().validNec[UserInputError]
+    }
+
+  def inputLockMatch(
+    testAddr:      LockAddress,
+    expectedAddr:  LockAddress,
+    testLabel:     String,
+    expectedLabel: String
+  ): ValidatedNec[UserInputError, Unit] =
+    Validated.condNec(
+      testAddr == expectedAddr,
+      (),
+      UserInputError(s"$testLabel does not correspond to $expectedLabel")
+    )
+
+  def positiveQuantity(testQuantity: Option[Int128], testLabel: String): ValidatedNec[UserInputError, Unit] =
+    Validated.condNec(
+      testQuantity.isEmpty || testQuantity.get > 0,
+      (),
+      UserInputError(s"$testLabel must be positive")
+    )
+
+  def validMintingSupply(
+    desiredQuantity: Int128,
+    seriesTokenOpt:  Option[Series],
+    testLabel:       String
+  ): ValidatedNec[UserInputError, Unit] =
+    seriesTokenOpt match {
+      case Some(Series(_, availableQuantity, Some(tokenSupply), _, _, _)) =>
+        if (desiredQuantity % tokenSupply != 0)
+          UserInputError(s"$testLabel must be a multiple of token supply").invalidNec[Unit]
+        else if (desiredQuantity > availableQuantity * tokenSupply)
+          UserInputError(s"$testLabel must be less than total token supply available.").invalidNec[Unit]
+        else ().validNec[UserInputError]
+      case _ => ().validNec[UserInputError]
+    }
+
+  def fungibilityType(testType: Option[FungibilityType]): ValidatedNec[UserInputError, Unit] =
+    Validated.condNec(
+      testType.isEmpty || testType.get == FungibilityType.GROUP_AND_SERIES,
+      (),
+      UserInputError(s"Unsupported fungibility type. We currently only support GROUP_AND_SERIES")
+    )
+
+  object TransactionBuilder {
+
+    // TODO: implement
+    def validateTransferParams(): Either[NonEmptyChain[UserInputError], Unit] = ???
+
+    /**
+     * Validates the parameters for minting group and series constructor tokens
+     * If user parameters are invalid, return a UserInputError.
+     */
+    def validateConstructorMintingParams(
+      registrationTxo:        Txo,
+      registrationLockAddr:   LockAddress,
+      policyRegistrationUtxo: TransactionOutputAddress,
+      quantityToMint:         Int128
+    ): Either[NonEmptyChain[UserInputError], Unit] =
+      Chain(
+        txoAddressMatch(registrationTxo.outputAddress, policyRegistrationUtxo, "registrationTxo", "registrationUtxo"),
+        isLvls(registrationTxo.transactionOutput.value.value, "registrationUtxo"),
+        inputLockMatch(
+          registrationLockAddr,
+          registrationTxo.transactionOutput.address,
+          "registrationLock",
+          "registrationTxo"
+        ),
+        positiveQuantity(quantityToMint.some, "quantityToMint")
+      ).fold.toEither
+
+    /**
+     * Validates the parameters for minting asset tokens
+     * If user parameters are invalid, return a UserInputError.
+     */
+    def validateAssetMintingParams(
+      mintingStatement: AssetMintingStatement,
+      groupTxo:         Txo,
+      seriesTxo:        Txo,
+      groupLockAddr:    LockAddress,
+      seriesLockAddr:   LockAddress
+    ): Either[NonEmptyChain[UserInputError], Unit] =
+      Chain(
+        txoAddressMatch(groupTxo.outputAddress, mintingStatement.groupTokenUtxo, "groupTxo", "groupTokenUtxo"),
+        txoAddressMatch(seriesTxo.outputAddress, mintingStatement.seriesTokenUtxo, "seriesTxo", "seriesTokenUtxo"),
+        isGroup(groupTxo.transactionOutput.value.value, "groupTxo"),
+        isSeries(seriesTxo.transactionOutput.value.value, "seriesTxo"),
+        inputLockMatch(
+          groupLockAddr,
+          groupTxo.transactionOutput.address,
+          "groupLock",
+          "groupTxo"
+        ),
+        inputLockMatch(
+          seriesLockAddr,
+          seriesTxo.transactionOutput.address,
+          "seriesLock",
+          "seriesTxo"
+        ),
+        positiveQuantity(mintingStatement.quantity.some, "quantity to mint"),
+        positiveQuantity(
+          seriesTxo.transactionOutput.value.value.series.map(_.quantity),
+          "quantity of input series constructor tokens"
+        ),
+        validMintingSupply(
+          mintingStatement.quantity,
+          seriesTxo.transactionOutput.value.value.series,
+          "quantity to mint"
+        ),
+        fixedSeriesMatch(
+          groupTxo.transactionOutput.value.value.group.flatMap(_.fixedSeries),
+          seriesTxo.transactionOutput.value.value.series.map(_.seriesId)
+        ),
+        fungibilityType(seriesTxo.transactionOutput.value.value.series.map(_.fungibility))
+      ).fold.toEither
+
+  }
+}
